Added almost all business logic and database realization. But not tested

+++ src/main/java/com/example/bankcards/dto/CardBalanceResponse.java
@@ -1,3 +1,3 @@
 package com.example.bankcards.dto;
 
-public record CardBalanceResponse(long balance) {}
+public record CardBalanceResponse(double balance) {}
