+++ src/main/java/com/example/bankcards/dto/CardInfoResponse.java
@@ -2,4 +2,4 @@ package com.example.bankcards.dto;
 
 import com.example.bankcards.entity.Card;
 
-public record CardInfoResponse(String maskedNumber, String date, Card.Status status, Long balance, String owner) {};
+public record CardInfoResponse(String maskedNumber, String date, Card.Status status, double balance, String owner) {};
